dont standardize qualities for everything
